server/requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
sqlalchemy
openai
python-dotenv